refactor: split CollapsibleList into Table & Row
